# Local span perf improvements