--- a/opentelemetry/src/global/trace.rs
+++ b/opentelemetry/src/global/trace.rs
@@ -21,7 +21,7 @@ impl trace::Span for BoxedSpan {
     /// keys"](https://github.com/open-telemetry/opentelemetry-specification/tree/v0.5.0/specification/trace/semantic_conventions/README.md)
     /// which have prescribed semantic meanings.
     fn add_event_with_timestamp(
-        &self,
+        &mut self,
         name: String,
         timestamp: SystemTime,
         attributes: Vec<KeyValue>,
@@ -45,23 +45,23 @@ impl trace::Span for BoxedSpan {
     /// Note that the OpenTelemetry project documents certain ["standard
     /// attributes"](https://github.com/open-telemetry/opentelemetry-specification/tree/v0.5.0/specification/trace/semantic_conventions/README.md)
     /// that have prescribed semantic meanings.
-    fn set_attribute(&self, attribute: KeyValue) {
+    fn set_attribute(&mut self, attribute: KeyValue) {
         self.0.set_attribute(attribute)
     }
 
     /// Sets the status of the `Span`. If used, this will override the default `Span`
     /// status, which is `Unset`.
-    fn set_status(&self, code: trace::StatusCode, message: String) {
+    fn set_status(&mut self, code: trace::StatusCode, message: String) {
         self.0.set_status(code, message)
     }
 
     /// Updates the `Span`'s name.
-    fn update_name(&self, new_name: String) {
+    fn update_name(&mut self, new_name: String) {
         self.0.update_name(new_name)
     }
 
     /// Finishes the span with given timestamp.
-    fn end_with_timestamp(&self, timestamp: SystemTime) {
+    fn end_with_timestamp(&mut self, timestamp: SystemTime) {
         self.0.end_with_timestamp(timestamp);
     }
 }
@@ -128,7 +128,7 @@ pub trait GenericTracer: fmt::Debug + 'static {
 
 impl<S, T> GenericTracer for T
 where
-    S: trace::Span + Send + Sync,
+    S: trace::Span + Send + Sync + 'static,
     T: trace::Tracer<Span = S>,
 {
     /// Create a new invalid span for use in cases where there are no active spans.
@@ -165,7 +165,7 @@ pub trait GenericTracerProvider: fmt::Debug + 'static {
 
 impl<S, T, P> GenericTracerProvider for P
 where
-    S: trace::Span + Send + Sync,
+    S: trace::Span + Send + Sync + 'static,
     T: trace::Tracer<Span = S> + Send + Sync,
     P: trace::TracerProvider<Tracer = T>,
 {
@@ -193,7 +193,7 @@ impl GlobalTracerProvider {
     /// Create a new GlobalTracerProvider instance from a struct that implements `TracerProvider`.
     fn new<P, T, S>(provider: P) -> Self
     where
-        S: trace::Span + Send + Sync,
+        S: trace::Span + Send + Sync + 'static,
         T: trace::Tracer<Span = S> + Send + Sync,
         P: trace::TracerProvider<Tracer = T> + Send + Sync,
     {
@@ -258,7 +258,7 @@ pub fn tracer_with_version(name: &'static str, version: &'static str) -> BoxedTr
 /// [`TracerProvider`]: crate::trace::TracerProvider
 pub fn set_tracer_provider<P, T, S>(new_provider: P) -> GlobalTracerProvider
 where
-    S: trace::Span + Send + Sync,
+    S: trace::Span + Send + Sync + 'static,
     T: trace::Tracer<Span = S> + Send + Sync,
     P: trace::TracerProvider<Tracer = T> + Send + Sync,
 {