--- a/opentelemetry/src/sdk/trace/span.rs
+++ b/opentelemetry/src/sdk/trace/span.rs
@@ -10,20 +10,14 @@
 //! These cannot be changed after the `Span`'s end time has been set.
 use crate::trace::{Event, SpanContext, SpanId, SpanKind, StatusCode};
 use crate::{sdk, trace, KeyValue};
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 use std::time::SystemTime;
 
 /// Single operation within a trace.
-#[derive(Clone, Debug)]
-pub struct Span {
-    inner: Arc<SpanInner>,
-}
-
-/// Inner data, processed and exported on end
 #[derive(Debug)]
-struct SpanInner {
+pub struct Span {
     span_context: SpanContext,
-    data: Option<Mutex<Option<SpanData>>>,
+    data: Option<SpanData>,
     tracer: sdk::trace::Tracer,
 }
 
@@ -49,8 +43,6 @@ pub(crate) struct SpanData {
     pub(crate) status_code: StatusCode,
     /// Span status message
     pub(crate) status_message: String,
-    /// Resource contains attributes representing an entity that produced this span.
-    pub(crate) resource: Arc<sdk::Resource>,
 }
 
 impl Span {
@@ -60,25 +52,18 @@ impl Span {
         tracer: sdk::trace::Tracer,
     ) -> Self {
         Span {
-            inner: Arc::new(SpanInner {
-                span_context,
-                data: data.map(|data| Mutex::new(Some(data))),
-                tracer,
-            }),
+            span_context,
+            data,
+            tracer,
         }
     }
 
     /// Operate on a mutable reference to span data
-    fn with_data<T, F>(&self, f: F) -> Option<T>
+    fn with_data<T, F>(&mut self, f: F) -> Option<T>
     where
         F: FnOnce(&mut SpanData) -> T,
     {
-        self.inner.data.as_ref().and_then(|inner| {
-            inner
-                .lock()
-                .ok()
-                .and_then(|mut span_data| span_data.as_mut().map(f))
-        })
+        self.data.as_mut().map(f)
     }
 }
 
@@ -89,7 +74,7 @@ impl crate::trace::Span for Span {
     /// keys"](https://github.com/open-telemetry/opentelemetry-specification/tree/v0.5.0/specification/trace/semantic_conventions/README.md)
     /// which have prescribed semantic meanings.
     fn add_event_with_timestamp(
-        &self,
+        &mut self,
         name: String,
         timestamp: SystemTime,
         attributes: Vec<KeyValue>,
@@ -102,19 +87,14 @@ impl crate::trace::Span for Span {
 
     /// Returns the `SpanContext` for the given `Span`.
     fn span_context(&self) -> &SpanContext {
-        &self.inner.span_context
+        &self.span_context
     }
 
     /// Returns true if this `Span` is recording information like events with the `add_event`
     /// operation, attributes using `set_attributes`, status with `set_status`, etc.
     /// Always returns false after span `end`.
     fn is_recording(&self) -> bool {
-        if let Some(data) = &self.inner.data {
-            if let Ok(span_data) = data.lock() {
-                return span_data.is_some();
-            }
-        }
-        false
+        self.data.is_some()
     }
 
     /// Sets a single `Attribute` where the attribute properties are passed as arguments.
@@ -122,7 +102,7 @@ impl crate::trace::Span for Span {
     /// Note that the OpenTelemetry project documents certain ["standard
     /// attributes"](https://github.com/open-telemetry/opentelemetry-specification/tree/v0.5.0/specification/trace/semantic_conventions/README.md)
     /// that have prescribed semantic meanings.
-    fn set_attribute(&self, attribute: KeyValue) {
+    fn set_attribute(&mut self, attribute: KeyValue) {
         self.with_data(|data| {
             data.attributes.insert(attribute);
         });
@@ -130,7 +110,7 @@ impl crate::trace::Span for Span {
 
     /// Sets the status of the `Span`. If used, this will override the default `Span`
     /// status, which is `Unset`. `message` MUST be ignored when the status is `OK` or `Unset`
-    fn set_status(&self, code: StatusCode, message: String) {
+    fn set_status(&mut self, code: StatusCode, message: String) {
         self.with_data(|data| {
             if code == StatusCode::Error {
                 data.status_message = message;
@@ -140,50 +120,49 @@ impl crate::trace::Span for Span {
     }
 
     /// Updates the `Span`'s name.
-    fn update_name(&self, new_name: String) {
+    fn update_name(&mut self, new_name: String) {
         self.with_data(|data| {
             data.name = new_name;
         });
     }
 
     /// Finishes the span with given timestamp.
-    fn end_with_timestamp(&self, timestamp: SystemTime) {
-        self.inner.ensure_ended_and_exported(Some(timestamp));
+    fn end_with_timestamp(&mut self, timestamp: SystemTime) {
+        self.ensure_ended_and_exported(Some(timestamp));
     }
 }
 
-impl SpanInner {
-    fn ensure_ended_and_exported(&self, timestamp: Option<SystemTime>) {
-        if let Some(data) = &self.data {
-            if let Ok(mut span_data) = data.lock().map(|mut data| data.take()) {
-                // Ensure end time is set via explicit end or implicitly on drop
-                if let Some(span_data) = span_data.as_mut() {
-                    if let Some(timestamp) = timestamp {
-                        span_data.end_time = timestamp;
-                    } else if span_data.end_time == span_data.start_time {
-                        span_data.end_time = crate::time::now();
-                    }
-                }
+impl Span {
+    fn ensure_ended_and_exported(&mut self, timestamp: Option<SystemTime>) {
+        if let Some(mut data) = self.data.take() {
+            // Ensure end time is set via explicit end or implicitly on drop
+            if let Some(timestamp) = timestamp {
+                data.end_time = timestamp;
+            } else if data.end_time == data.start_time {
+                data.end_time = crate::time::now();
+            }
 
-                // Notify each span processor that the span has ended
-                if let Some(provider) = self.tracer.provider() {
-                    let mut processors = provider.span_processors().iter().peekable();
-                    while let Some(processor) = processors.next() {
-                        let span_data = if processors.peek().is_none() {
-                            // last loop or single processor/exporter, move data
-                            span_data.take()
-                        } else {
-                            // clone so each exporter gets owned data
-                            span_data.clone()
-                        };
+            // Notify each span processor that the span has ended
+            if let Some(provider) = self.tracer.provider() {
+                let mut processors = provider.span_processors().iter().peekable();
+                let resource = provider.config().resource.clone();
+                let mut span_data = Some((data, resource));
+                while let Some(processor) = processors.next() {
+                    let span_data = if processors.peek().is_none() {
+                        // last loop or single processor/exporter, move data
+                        span_data.take()
+                    } else {
+                        // clone so each exporter gets owned data
+                        span_data.clone()
+                    };
 
-                        if let Some(span_data) = span_data {
-                            processor.on_end(build_export_data(
-                                span_data,
-                                self.span_context.clone(),
-                                &self.tracer,
-                            ));
-                        }
+                    if let Some((span_data, resource)) = span_data {
+                        processor.on_end(build_export_data(
+                            span_data,
+                            self.span_context.clone(),
+                            resource,
+                            &self.tracer,
+                        ));
                     }
                 }
             }
@@ -191,7 +170,7 @@ impl SpanInner {
     }
 }
 
-impl Drop for SpanInner {
+impl Drop for Span {
     /// Report span on inner drop
     fn drop(&mut self) {
         self.ensure_ended_and_exported(None);
@@ -201,6 +180,7 @@ impl Drop for SpanInner {
 fn build_export_data(
     data: SpanData,
     span_context: SpanContext,
+    resource: Arc<sdk::Resource>,
     tracer: &sdk::trace::Tracer,
 ) -> sdk::export::trace::SpanData {
     sdk::export::trace::SpanData {
@@ -215,7 +195,7 @@ fn build_export_data(
         links: data.links,
         status_code: data.status_code,
         status_message: data.status_message,
-        resource: data.resource,
+        resource,
         instrumentation_lib: *tracer.instrumentation_library(),
     }
 }
@@ -241,7 +221,6 @@ mod tests {
             links: sdk::trace::EvictedQueue::new(config.max_links_per_span),
             status_code: StatusCode::Unset,
             status_message: "".to_string(),
-            resource: config.resource.clone(),
         };
         (tracer, data)
     }
@@ -254,20 +233,20 @@ mod tests {
     #[test]
     fn create_span_without_data() {
         let (tracer, _) = init();
-        let span = Span::new(SpanContext::empty_context(), None, tracer);
+        let mut span = Span::new(SpanContext::empty_context(), None, tracer);
         span.with_data(|_data| panic!("there are data"));
     }
 
     #[test]
     fn create_span_with_data_mut() {
         let (tracer, data) = init();
-        let span = Span::new(SpanContext::empty_context(), Some(data.clone()), tracer);
+        let mut span = Span::new(SpanContext::empty_context(), Some(data.clone()), tracer);
         span.with_data(|d| assert_eq!(*d, data));
     }
 
     #[test]
     fn add_event() {
-        let span = create_span();
+        let mut span = create_span();
         let name = "some_event".to_string();
         let attributes = vec![KeyValue::new("k", "v")];
         span.add_event(name.clone(), attributes.clone());
@@ -283,7 +262,7 @@ mod tests {
 
     #[test]
     fn add_event_with_timestamp() {
-        let span = create_span();
+        let mut span = create_span();
         let name = "some_event".to_string();
         let attributes = vec![KeyValue::new("k", "v")];
         let timestamp = crate::time::now();
@@ -301,7 +280,7 @@ mod tests {
 
     #[test]
     fn record_exception() {
-        let span = create_span();
+        let mut span = create_span();
         let err = std::io::Error::from(std::io::ErrorKind::Other);
         span.record_exception(&err);
         span.with_data(|data| {
@@ -319,7 +298,7 @@ mod tests {
 
     #[test]
     fn record_exception_with_stacktrace() {
-        let span = create_span();
+        let mut span = create_span();
         let err = std::io::Error::from(std::io::ErrorKind::Other);
         let stacktrace = "stacktrace...".to_string();
         span.record_exception_with_stacktrace(&err, stacktrace.clone());
@@ -341,7 +320,7 @@ mod tests {
 
     #[test]
     fn set_attribute() {
-        let span = create_span();
+        let mut span = create_span();
         let attributes = KeyValue::new("k", "v");
         span.set_attribute(attributes.clone());
         span.with_data(|data| {
@@ -356,7 +335,7 @@ mod tests {
     #[test]
     fn set_status() {
         {
-            let span = create_span();
+            let mut span = create_span();
             let status = StatusCode::Ok;
             let message = "OK".to_string();
             span.set_status(status, message);
@@ -366,7 +345,7 @@ mod tests {
             });
         }
         {
-            let span = create_span();
+            let mut span = create_span();
             let status = StatusCode::Unset;
             let message = "OK".to_string();
             span.set_status(status, message);
@@ -376,7 +355,7 @@ mod tests {
             });
         }
         {
-            let span = create_span();
+            let mut span = create_span();
             let status = StatusCode::Error;
             let message = "Error".to_string();
             span.set_status(status, message);
@@ -389,7 +368,7 @@ mod tests {
 
     #[test]
     fn update_name() {
-        let span = create_span();
+        let mut span = create_span();
         let name = "new_name".to_string();
         span.update_name(name.clone());
         span.with_data(|data| {
@@ -399,13 +378,13 @@ mod tests {
 
     #[test]
     fn end() {
-        let span = create_span();
+        let mut span = create_span();
         span.end();
     }
 
     #[test]
     fn end_with_timestamp() {
-        let span = create_span();
+        let mut span = create_span();
         let timestamp = crate::time::now();
         span.end_with_timestamp(timestamp);
         span.with_data(|data| assert_eq!(data.end_time, timestamp));
@@ -413,21 +392,14 @@ mod tests {
 
     #[test]
     fn allows_to_get_span_context_after_end() {
-        let span = create_span();
+        let mut span = create_span();
         span.end();
         assert_eq!(span.span_context(), &SpanContext::empty_context());
     }
 
     #[test]
-    fn allows_to_get_span_context_after_clone_drop() {
-        let span = create_span();
-        drop(span.clone());
-        assert_eq!(span.span_context(), &SpanContext::empty_context());
-    }
-
-    #[test]
     fn end_only_once() {
-        let span = create_span();
+        let mut span = create_span();
         let timestamp = crate::time::now();
         span.end_with_timestamp(timestamp);
         span.end_with_timestamp(timestamp.checked_add(Duration::from_secs(10)).unwrap());
@@ -436,7 +408,7 @@ mod tests {
 
     #[test]
     fn noop_after_end() {
-        let span = create_span();
+        let mut span = create_span();
         let initial = span.with_data(|data| data.clone()).unwrap();
         span.end();
         span.add_event("some_event".to_string(), vec![KeyValue::new("k", "v")]);
@@ -468,7 +440,7 @@ mod tests {
 
     #[test]
     fn is_recording_false_after_end() {
-        let span = create_span();
+        let mut span = create_span();
         span.end();
         assert!(!span.is_recording());
     }