--- a/opentelemetry/src/testing/trace.rs
+++ b/opentelemetry/src/testing/trace.rs
@@ -19,7 +19,7 @@ pub struct TestSpan(pub SpanContext);
 
 impl Span for TestSpan {
     fn add_event_with_timestamp(
-        &self,
+        &mut self,
         _name: String,
         _timestamp: std::time::SystemTime,
         _attributes: Vec<KeyValue>,
@@ -31,10 +31,10 @@ impl Span for TestSpan {
     fn is_recording(&self) -> bool {
         false
     }
-    fn set_attribute(&self, _attribute: KeyValue) {}
-    fn set_status(&self, _code: StatusCode, _message: String) {}
-    fn update_name(&self, _new_name: String) {}
-    fn end_with_timestamp(&self, _timestamp: std::time::SystemTime) {}
+    fn set_attribute(&mut self, _attribute: KeyValue) {}
+    fn set_status(&mut self, _code: StatusCode, _message: String) {}
+    fn update_name(&mut self, _new_name: String) {}
+    fn end_with_timestamp(&mut self, _timestamp: std::time::SystemTime) {}
 }
 
 pub fn new_test_export_span_data() -> SpanData {